troubleshooting navbar issue:
Navbar is a different size on the about page...?
it also does some other strange things like when you are on the about page it will be smaller
text than any other page. goes back to normal on refresh...?
so as i learned more i realized i should have the nav and footer appear in an element that wraps the entire appear
I had it in _document.js first but the same issue or perhaps the cause?
Removing it here and will work from there. so now now footer and nav appear in the app anywhere at the point
Also removed whitespace Header component